feat: cleaned up electron section
