+++ src/main/factories/menu.ts
@@ -1,99 +1,39 @@
-import { app, shell, Menu, MenuItem, MenuItemConstructorOptions, BrowserWindow } from 'electron';
+import { app, shell, Menu, MenuItem, MenuItemConstructorOptions } from 'electron';
 
-import { ENVIRONMENT, PLATFORM } from '../shared/constants';
+import { PLATFORM } from '../shared/constants/platform';
+import { ENVIRONMENT } from '../shared/constants/environment';
 
-export function buildMenuTemplate(window: BrowserWindow): Menu {
+export function buildMenuTemplate(): Menu {
   const template: MenuItemConstructorOptions[] | MenuItem[] = [
-    {
-      label: 'TEST',
-      submenu: [
-        {
-          click: () => window.webContents.send('protocol-open', 'test data  '),
-          label: 'send url',
-        },
-      ],
-    },
     {
       label: 'Edit',
       accelerator: 'e',
       submenu: [
-        {
-          role: 'undo',
-          label: 'Undo',
-        },
-        {
-          role: 'redo',
-          label: 'Redo',
-        },
+        { role: 'undo', label: 'Undo' },
+        { role: 'redo', label: 'Redo' },
         { type: 'separator' },
-        {
-          role: 'cut',
-          label: 'Cut',
-        },
-        {
-          role: 'copy',
-          label: 'Copy',
-        },
-        {
-          role: 'paste',
-          label: 'Paste',
-        },
-        {
-          role: 'pasteAndMatchStyle',
-          label: 'Paste and Match Style',
-        },
-        {
-          role: 'delete',
-          label: 'Delete',
-        },
-        {
-          role: 'selectAll',
-          label: 'Select All',
-        },
+        { role: 'cut', label: 'Cut' },
+        { role: 'copy', label: 'Copy' },
+        { role: 'paste', label: 'Paste' },
+        { role: 'pasteAndMatchStyle', label: 'Paste and Match Style' },
+        { role: 'delete', label: 'Delete' },
+        { role: 'selectAll', label: 'Select All' },
       ],
     },
     {
       label: 'View',
       accelerator: 'V',
       submenu: [
-        {
-          role: 'reload',
-        },
-        {
-          role: 'forceReload',
-        },
+        { role: 'reload' },
+        { role: 'forceReload' },
         { type: 'separator' },
-        {
-          role: 'resetZoom',
-          accelerator: 'CmdOrCtrl+Num0',
-          visible: false,
-        },
-        {
-          role: 'zoomIn',
-          accelerator: 'CmdOrCtrl+NumAdd',
-          visible: false,
-        },
-        {
-          role: 'zoomOut',
-          accelerator: 'CmdOrCtrl+NumSub',
-          visible: false,
-        },
-        {
-          role: 'resetZoom',
-          label: 'Actual Size',
-        },
-        {
-          role: 'zoomIn',
-          label: 'Zoom In',
-        },
-        {
-          role: 'zoomOut',
-          label: 'Zoom Out',
-        },
-        {
-          role: 'togglefullscreen',
-          label: 'Toggle Full Screen',
-        },
+        { role: 'resetZoom', accelerator: 'CmdOrCtrl+Num0', visible: false },
+        { role: 'zoomIn', accelerator: 'CmdOrCtrl+NumAdd', visible: false },
+        { role: 'zoomOut', accelerator: 'CmdOrCtrl+NumSub', visible: false },
+        { role: 'resetZoom', label: 'Actual Size' },
+        { role: 'zoomIn', label: 'Zoom In' },
+        { role: 'zoomOut', label: 'Zoom Out' },
+        { role: 'togglefullscreen', label: 'Toggle Full Screen' },
       ],
     },
     {
@@ -101,14 +41,8 @@ export function buildMenuTemplate(window: BrowserWindow): Menu {
       accelerator: 'w',
       role: 'window',
       submenu: [
-        {
-          role: 'minimize',
-          label: 'Minimize',
-        },
-        {
-          role: 'close',
-          label: 'Close',
-        },
+        { role: 'minimize', label: 'Minimize' },
+        { role: 'close', label: 'Close' },
       ],
     },
     {
@@ -118,7 +52,7 @@ export function buildMenuTemplate(window: BrowserWindow): Menu {
       submenu: [
         {
           label: 'Nova Spektr Help',
-          click(): void {
+          click() {
             shell.openExternal('https://docs.novaspektr.io/');
           },
         },
@@ -136,66 +70,35 @@ export function buildMenuTemplate(window: BrowserWindow): Menu {
     }
   }
 
-  // macOS has specific menu conventions...
+  // macOS has specific menu conventions
   if (PLATFORM.IS_MAC) {
     template.unshift({
       // first macOS menu is the name of the app
       role: 'appMenu',
       label: app.name,
       submenu: [
-        {
-          role: 'about',
-          label: 'About Nova Spektr',
-        },
+        { role: 'about', label: 'About Nova Spektr' },
         { type: 'separator' },
-        {
-          role: 'hide',
-          label: 'Hide',
-        },
-        {
-          role: 'hideOthers',
-          label: 'Hide Others',
-        },
-        {
-          role: 'unhide',
-          label: 'Unhide',
-        },
+        { role: 'hide', label: 'Hide' },
+        { role: 'hideOthers', label: 'Hide Others' },
+        { role: 'unhide', label: 'Unhide' },
         { type: 'separator' },
-        {
-          role: 'quit',
-          label: 'Quit',
-        },
+        { role: 'quit', label: 'Quit' },
       ],
     });
 
-    // Window menu.
+    // Window menu
     // This also has specific functionality on macOS
     template[3].submenu = [
-      {
-        label: 'Close',
-        accelerator: 'CmdOrCtrl+W',
-        role: 'close',
-      },
-      {
-        label: 'Minimize',
-        accelerator: 'CmdOrCtrl+M',
-        role: 'minimize',
-      },
-      {
-        label: 'Zoom',
-        role: 'zoom',
-      },
+      { label: 'Close', accelerator: 'CmdOrCtrl+W', role: 'close' },
+      { label: 'Minimize', accelerator: 'CmdOrCtrl+M', role: 'minimize' },
+      { label: 'Zoom', role: 'zoom' },
     ];
   } else {
     template.unshift({
       label: 'File',
       accelerator: 'f',
-      submenu: [
-        {
-          role: 'quit',
-          label: 'Quit',
-        },
-      ],
+      submenu: [{ role: 'quit', label: 'Quit' }],
     });
   }
