--- src/main/shared/constants/index.ts
@@ -1,4 +1,0 @@
-export { ENVIRONMENT } from './environment';
-export { PLATFORM } from './platform';
-export { IPC } from './ipc';
-export { AUTO_UPDATE_ENABLED } from './electron-store';
